+++ zest/releaser/git.py
@@ -4,7 +4,7 @@ import os
 import os.path
 import sys
 
-from zest.releaser.utils import system
+from zest.releaser.utils import execute_command
 from zest.releaser.vcs import BaseVersionControl
 
 logger = logging.getLogger(__name__)
@@ -35,7 +35,7 @@ class Git(BaseVersionControl):
         return dir_name
 
     def available_tags(self):
-        tag_info = system('git tag')
+        tag_info = execute_command('git tag')
         tags = [line for line in tag_info.split('\n') if line]
         logger.debug("Available tags: %r", tags)
         return tags
@@ -47,11 +47,11 @@ class Git(BaseVersionControl):
         cwd = os.getcwd()
         os.chdir(temp)
         cmd = 'git clone %s %s' % (self.workingdir, 'gitclone')
-        logger.debug(system(cmd))
+        logger.debug(execute_command(cmd))
         clonedir = os.path.join(temp, 'gitclone')
         os.chdir(clonedir)
         cmd = 'git submodule update --init --recursive'
-        logger.debug(system(cmd))
+        logger.debug(execute_command(cmd))
         os.chdir(cwd)
         return clonedir
 
@@ -124,14 +124,14 @@ class Git(BaseVersionControl):
     def is_clean_checkout(self):
         """Is this a clean checkout?
         """
-        head = system('git symbolic-ref --quiet HEAD')
+        head = execute_command('git symbolic-ref --quiet HEAD')
         # This returns something like 'refs/heads/maurits-warn-on-tag'
         # or nothing.  Nothing would be bad as that indicates a
         # detached head: likely a tag checkout
         if not head:
             # Greetings from Nearly Headless Nick.
             return False
-        if system('git status --short --untracked-files=no'):
+        if execute_command('git status --short --untracked-files=no'):
             # Uncommitted changes in files that are tracked.
             return False
         return True
@@ -142,4 +142,4 @@ class Git(BaseVersionControl):
 
     def list_files(self):
         """List files in version control."""
-        return system('git ls-tree -r HEAD --name-only').splitlines()
+        return execute_command('git ls-tree -r HEAD --name-only').splitlines()