+++ CHANGES.rst
@@ -4,6 +4,14 @@ Changelog for zest.releaser
 4.0 (unreleased)
 ----------------
 
+- Try not to treat warnings as errors.
+  [maurits]
+
+- Allow retrying some commands when there is an error.  Currently only
+  for commands that talk to PyPI or another package index.  We ask the
+  user if she wants to retry: Yes, no, quit.
+  [maurits]
+
 - Added support for twine_.  If the ``twine`` command is available, it
   is used for uploading to PyPI.  It is installed automatically if you
   use the ``zest.releaser[recommended]`` extra.  Note that if the